SETTING THE INITIAL STATE
the variable movementstate has been set to closing in the setup function. This combined with the code logic sets the system up as shutter closed on power up. There is a problem with this though. If the system fails with the shutter open, the power on reset only allows for an OPEN command (OS#) which would destroy the shutter chain drive.